test(e2e): Add E2E tests (#75)
This PR adds basic E2E or Snapshot tests to our repo in a dedicated package. The high-level idea is to add a set of input JS files, a plugin config and a set of reference files per test scenario. For a more detailed explanation, take a look at the README file. In addition, it adds one test scenario which tests a very basic sourcemaps upload configuration.

Here's what these tests do:

Per test scenario, per bundler:
1. Delete the release from Sentry.io (sentry-sdks org) if it is already present
2. Bundle the input files and run the Sentry Bundler Plugin which uploads the files to Sentry
3. Download the release from Sentry 
4. Compare downloaded files against reference files. They must match exactly for the test to pass. 

As of right now, it is arguable if these tests should have their own package or be in the integration-tests package. I opted for the former because there are subtle differences and if we chose to improve them (e.g. do the Verdaccio approach from the JS SDK or add recipes) the structure and components will differ more.