ref(core): Remove customHeader conversion in internal options (#98)
Previously, we coverted the user-facing `customHeader` option, typed as string to an internal option typed `Record<string, string>`. This was handy for adding the header to our outgoing API requests. Since we don't do these ourselves anymore but rely on Sentry CLI, it's just easier to pass this option through to the CLI which also expects the string representation.