feat(core): Add plugin options as tags (#113)
Add a subset of our plugin options as tags to Sentry events. Obviously, we cannot send all options but we have to be careful to only send insensitive data, such as if an optional step (e.g. `deploy`) was selected or not. 

Going forward, we can use this data to evaluate which options are used and for debugging issues. 

Also add some tests